Renamed folder components/order to component/orders
